# Change bound checking in SSE/AVX intrinsics to avoid pointer overflow