--- a/src/Microsoft.ML.CpuMath/SseIntrinsics.cs
+++ b/src/Microsoft.ML.CpuMath/SseIntrinsics.cs
@@ -26,9 +26,6 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             Sse.StaticCast<int, float>(Sse2.SetAllVector128(0x7FFFFFFF)) :
             Sse.SetAllVector128(BitConverter.Int32BitsToSingle(0x7FFFFFFF));
 
-        // The count of 32-bit floats in Vector128<T>
-        internal const int Vector128SingleElementCount = 4;
-
         // The count of bytes in Vector128<T>, corresponding to _cbAlign in AlignedArray
         private const int Vector128Alignment = 16;
 
@@ -416,7 +413,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             fixed (float* pdst = dst)
             {
                 Vector128<float> scalarVector = Sse.SetAllVector128(scalar);
-                int count = Math.DivRem(dst.Length, Vector128SingleElementCount, out int remainder);
+                int count = Math.DivRem(dst.Length, 4, out int remainder);
                 float* pDstCurrent = pdst;
 
                 for (int i = 0; i < count; i++)
@@ -425,7 +422,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
                     dstVector = Sse.Add(dstVector, scalarVector);
                     Sse.Store(pDstCurrent, dstVector);
 
-                    pDstCurrent += Vector128SingleElementCount;
+                    pDstCurrent += 4;
                 }
 
                 for (int i = 0; i < remainder; i++)