@eerhardt 2018-09-05
I’m not sure I follow the reasoning. These are pointer operations. So the concern is that we are less than 4 or 8 elements away from the end of the memory?  Would the OS ever let us get that close?