@ahsonkhan 2018-09-05
> Another question I have is: would pDstCurrent + 8 OR 4 ever have the possibility to result in integer overflow

Imo, given these are public APIs that anyone can call (with potentially invalid inputs), the inputs should be verified within the method body. Being explicit about assertions like src length == dst length would be good (and also act as self-documentation).

> According to my knowledge, pEnd is initialized as pDstCurrent + count, and there are Contract.Asserts in the wrapper class to check that count does not exceed the original array length. I'm not sure, and am open to any PR comments and advice.

I am not too familiar with the code base here, but don't contract.asserts run in debug mode only? Are these checks unnecessary in release?

> Both SSE and AVX implementations are slower by 10-20% after this change.

How about something like the following:
```C#
int i = 0;
for (; i < src.Length - 8; i += 8)
{
    Vector256<float> srcVector = Avx.LoadVector256(pSrcCurrent);
    Vector256<float> dstVector = Avx.LoadVector256(pDstCurrent);

    result256 = Avx.Add(result256, Avx.Multiply(srcVector, dstVector));

    pSrcCurrent += 8;
    pDstCurrent += 8;
}

if (src.Length - i <= 4)
{
    ...
    i += 4;
}
while (i < src.Length)
{
    ...
    i++;
}
```

The number of assembly instruction is essentially identical here (saves a `lea`, costs an extra `add`): https://www.diffchecker.com/MywBeoFH (before in red, after in green)

![image](https://user-images.githubusercontent.com/6527137/45066006-0a930380-b071-11e8-93c0-2ecdaf442d2e.png)

Just my two cents. I would leave it up to others who have more context in this space to validate.

> Would the OS ever let us get that close?

No, it won't. I discussed this with @GrabYourPitchforks, and its a general coding guideline to avoid arithmetic overflows like this (on the, albeit unlikely, chance the OS behavior changes in the future).