@ahsonkhan 2018-09-05
> Following your logic, would changing while (pCurrent + 8 OR 4 <= pEnd) into while (pCurrent <= pEnd - 8 OR 4) have a similar effect?

That has a similar concern (but with underflow). If src.Length < 8, and in the unlikely chance that it starts at a memory address < 8, then `(pCurrent <= pEnd - 8)` would be true, even though it shouldn't. That is because comparisons are done as if they are unsigned integers.

https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/unsafe-code-pointers/pointer-comparison
> The comparison operators compare the addresses of the two operands as if they are unsigned integers.

**For example:**
```C#
// Assume:
src.Length = 6;
pCurrent = 4;
pEnd = pCurrent + src.Length // 4 + 6 * 4 = 28;

if (pCurrent <= pEnd - 8)
{
   // 4 <= 28 - (8 * 4)
   // 4 <= 28 - 32
   // 4 <= (ulong)-4
   // We don't expect to be here, but we will since -4 as a ulong is a really large number.
}
```