@GrabYourPitchforks 2018-09-06
> So the concern is that we are less than 4 or 8 elements away from the end of the memory? Would the OS ever let us get that close?

@ahsonkhan and I spoke about this at length yesterday.

With _current_ operating systems, no. But I don't know what OSes we'll be running on in 10 years. Maybe we'll have a fully managed OS like Singularity, and maybe it'll give us access to the full range of addressable memory. I can't predict the future, and I don't want to chance having to chase down logic errors in this code in ten years' time if there's a theoretical overflow condition that we can identify and address now.