@GrabYourPitchforks 2018-09-06
FWIW, there is an exception that both C# and C allow when comparing pointers. The element just past the end of the array is always addressable, and it's always guaranteed to compare greater than the address of any element in the array.

That is:

```c
T* ptr; // = pointer to first element in array
size_t numElements; // total number of elements in the array

for (size_t i = 0; i < numElements; i++) {
    assert(&ptr[i] < &ptr[numElements], "This is guaranteed by the language.");
}
```

A corollary to this is that `T* end = ptr + numElements` is guaranteed _not_ to integer overflow (so the array can't be placed at the very, very end of addressable memory), but `T* end = ptr + numElements + 1` has no such guarantee.