@eerhardt 2018-09-06
Ah, ok.  I did some more reading, looked at this code again and I understand the concern. The array may be at the very end of memory space. And when `pDstCurrent` is less than 8 elements from the end, then `while (pDstCurrent + 8 <= pDstEnd)` will go out of bounds, and potentially overflow. And if it overflows, it will wrap around making condition `true`.

I think it makes sense using the `for (; i < src.Length - 8; i += 8)` pattern that @ahsonkhan proposes above.