@briancylui 2018-09-06
Thanks for all the comments.  After some reading, I agree that the problem is potential illegal memory access due to incrementing the `pCurrent` pointer out of bounds of the `Span<float>` object.  Accessing out-of-bound object can already cause security concerns, and it would be even worse if the pointer got incremented into illegal memory.  I should probably change the title of this PR to "avoid pointer overflow" or "avoid illegal memory access" to avoid any possible confusion, and look forward to implementing this fix tomorrow.

A very good takeaway from @ahsonkhan and @GrabYourPitchforks's comments is that we may not want to increment/decrement any pointer for bound checking, and I think it is also a very good opportunity to follow through on @eerhardt and @tannergooding's suggestion, which will pave the way for implementing double-computing.  I will spend some time tomorrow to look into how to implement these fixes concretely.

Thank you very much for all the comments!