@briancylui 2018-09-07
Pushed a new commit and deleted the previous commit to show you all a sample of the changes I will make globally.

In the new commit, only the `AddScalarU` intrinsic has been changed in its SSE and AVX implementations.  Once you think that the changes look fine to you, or you may some suggestions to edit them, the changes can be made on a global scale to the rest of the intrinsics.

The performance test results are shown below:

## After the change:
``` ini
BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.228 (1803/April2018Update/Redstone4)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host] : .NET Core 3.0.0-preview1-26710-03 (CoreCLR 4.6.26710.05, CoreFX 4.6.26708.04), 64bit RyuJIT

Toolchain=InProcessToolchain
```
|                   Type |     Method |     Mean |     Error |    StdDev |   Median |
|----------------------- |----------- |---------:|----------:|----------:|---------:|
|    AvxPerformanceTests | AddScalarU | 142.9 us | 2.8338 us | 7.0045 us | 140.1 us |
| NativePerformanceTests | AddScalarU | 191.2 us | 2.3916 us | 1.8672 us | 190.9 us |
|    SsePerformanceTests | AddScalarU | 172.7 us | 0.7552 us | 0.5896 us | 172.6 us |

## Before the change:
``` ini
BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.228 (1803/April2018Update/Redstone4)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host] : .NET Core 3.0.0-preview1-26710-03 (CoreCLR 4.6.26710.05, CoreFX 4.6.26708.04), 64bit RyuJIT

Toolchain=InProcessToolchain
```
|                   Type |     Method |     Mean |    Error |    StdDev |   Median |
|----------------------- |----------- |---------:|---------:|----------:|---------:|
|    AvxPerformanceTests | AddScalarU | 141.8 us | 1.623 us |  1.439 us | 141.5 us |
| NativePerformanceTests | AddScalarU | 191.8 us | 4.514 us | 10.901 us | 186.5 us |
|    SsePerformanceTests | AddScalarU | 180.3 us | 3.653 us |  7.125 us | 177.1 us |

The performance results are comparable after the change: SEE got faster by a bit, while AVX got slower by just a bit.  The second use of `Math.DivRem` for `countSse` and `remainderSse` may have made the AVX implementation a bit slower.

Look forward to all your comments.  Since tomorrow is the last day of my internship, and I have other work items to finish before I go, I don't think I can finish this PR by tomorrow, but I will keep my branch in my fork up so that other contributors can use your previous commits.  While I will leave this PR open for the moment, please feel free to close it whenever you feel appropriate.  Thank you.